pyproject.toml
[project]
name = "egoerrorbench-a2a"
version = "0.1.0"
description = "EgoErrorVQA evaluation framework using A2A protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "qwen-vl-utils>=0.0.1",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.0.0",
    "av>=12.0.0",
    "numpy>=1.24.0",
    "accelerate>=0.20.0",
    "decord>=0.6.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
